refactor: moved like function to component
resolves #363
