+++ src/routes/lessons/+page.svelte
@@ -1,19 +1,11 @@
 <script>
-  import { Menu, MakePlaylist, OwnPlaylists, AllStories, LikedPlaylists, SuggestedPlaylists   } from '$lib/index';
+  import { Menu, MakePlaylist, AllStories, PlaylistLike   } from '$lib/index';
 
   let currentPage = "lessons";
 
   /** @type {import('./$types').PageData} */
   export let data;
 
-  function handleLikeToggle(event) {
-    const { playlistId, isLiked } = event.detail;
-    data.playlists = data.playlists.map(playlist =>
-      playlist.id === playlistId
-        ? { ...playlist, isLiked: isLiked }
-        : playlist
-    );
-  }
 </script>
 
 <main>
@@ -26,39 +18,38 @@
     </ul>
   </header>
 
-  <OwnPlaylists {data} on:likeToggle={handleLikeToggle} />
+  <!-- {PLaylistLike} includes {Own playlists, Liked playlist & Suggested playlists} -->
+  <PlaylistLike {data} />
   <MakePlaylist {data} />
   <AllStories {data} />
-  <LikedPlaylists {data} on:likeToggle={handleLikeToggle} />
-  <SuggestedPlaylists {data} on:likeToggle={handleLikeToggle} />
 </main>
 
 <style>
 
-  main {
-    height: 100%;
-    color: var(--color-text-light);
-    background-image: var(--bg-image-purple);
-    display: flex;
-    flex-direction: column;
-    padding-bottom: 5em;
-  }
+main {
+  height: 100%;
+  color: var(--color-text-light);
+  background-image: var(--bg-image-purple);
+  display: flex;
+  flex-direction: column;
+  padding-bottom: 5em;
+}
 
-  header {
-    margin-bottom: var(--space-lg);
-    display: flex;
-    flex-direction: column;
+header {
+  margin-bottom: var(--space-lg);
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+h1 {
+  margin-top: var(--space-lg);
+  margin-bottom: var(--space-md);
+}
+
+@media (min-width: 600px) {
+  main {
     align-items: center;
   }
-
-  h1 {
-    margin-top: var(--space-lg);
-    margin-bottom: var(--space-md);
-  }
-
-  @media (min-width: 600px) {
-    main {
-      align-items: center;
-    }
-  }
+}
 </style>
